Add role-based access control for inventory, orders, and payments

+++ src/main/java/com/example/animeecommercebackend/Service/Impl/OrderServiceImpl.java
@@ -6,6 +6,7 @@ import com.example.animeecommercebackend.Entity.Order;
 import com.example.animeecommercebackend.Exception.ResourceNotFoundException;
 import com.example.animeecommercebackend.Mapper.OrderMapper;
 import com.example.animeecommercebackend.Repository.OrderRepository;
+import com.example.animeecommercebackend.Service.CurrentUserService;
 import com.example.animeecommercebackend.Service.OrderService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -16,19 +17,38 @@ import java.util.List;
 @RequiredArgsConstructor
 public class OrderServiceImpl implements OrderService {
     private final OrderRepository orderRepository;
+    private final CurrentUserService currentUserService;
+
+
     @Override
     public OrderResponseDto createOrder(OrderRequestDto dto) {
         Order order = OrderMapper.toEntity(dto);
 
+        order.setUser(currentUserService.getCurrentUser());
         Order saved = orderRepository.save(order);
 
         return OrderMapper.toResponse(saved);
+    }
+    @Override
+    public List<OrderResponseDto> getMyOrders() {
+        Long currentUserId = currentUserService.getCurrentUser().getId();
+
+        return orderRepository.findByUserId(currentUserId)
+                .stream()
+                .map(OrderMapper::toResponse)
+                .toList();
     }
 
     @Override
     public OrderResponseDto getOrderById(Long id) {
         Order order = orderRepository.findById(id).orElseThrow(() ->
                 new ResourceNotFoundException("Order Not Found!"));
+
+        Long currentUserId = currentUserService.getCurrentUser().getId();
+
+        if(!order.getId().equals(currentUserId)){
+            throw new ResourceNotFoundException("Order Not found");
+        }
         return OrderMapper.toResponse(order);
     }
 
@@ -51,6 +71,11 @@ public class OrderServiceImpl implements OrderService {
         Order order = orderRepository.findById(id).orElseThrow(() ->
                 new ResourceNotFoundException("Order Not Found!"));
 
+        Long currentUserId = currentUserService.getCurrentUser().getId();
+
+        if(!order.getUser().getId().equals(currentUserId)){
+            throw new ResourceNotFoundException("Order Not Found!!");
+        }
         order.setShippingAddress(dto.getShippingAddress());
 
         Order saved = orderRepository.save(order);
@@ -61,6 +86,12 @@ public class OrderServiceImpl implements OrderService {
     public void deleteOrder(Long id) {
         Order order = orderRepository.findById(id).orElseThrow(() ->
                 new ResourceNotFoundException("Order Not Found!"));
+
+        Long currentUserId = currentUserService.getCurrentUser().getId();
+
+        if(!order.getUser().getId().equals(currentUserId)){
+            throw new ResourceNotFoundException("Order Not Found");
+        }
 
         orderRepository.delete(order);
     }
